requirements.txt
tensorflow==2.9.0
pillow==9.2.0
numpy==1.23.1
